+++ src/editor/breadcrumb.ts
@@ -1,5 +1,5 @@
 import { editorColorDef } from '../color';
 
-export default function setBreadcrumb(data: EditorColorMap) {
+export default function setBreadcrumb(data: ColorMap) {
   data.set('breadcrumb.background', editorColorDef.bgDark);
 }