+++ src/editor/general.ts
@@ -1,5 +1,5 @@
 import { editorColorDef } from '../color';
 
-export default function editorGeneralSetting(data: EditorColorMap) {
+export default function editorGeneralSetting(data: ColorMap) {
   data.set('foreground', editorColorDef.fgColor);
 }