+++ src/editor/sideBar.ts
@@ -1,6 +1,6 @@
 import { editorColorDef } from '../color';
 
-export default function setSidebarColor(data: EditorColorMap) {
+export default function setSidebarColor(data: ColorMap) {
   data.set('sideBar.background', editorColorDef.bgColor);
   data.set('sideBarTitle.foreground', editorColorDef.fgDark);
   data.set('sideBarSectionHeader.background', editorColorDef.bgDark);