+++ src/editor/statusBar.ts
@@ -1,6 +1,6 @@
 import { editorColorDef } from '../color';
 
-export default function setStatusBar(data: EditorColorMap) {
+export default function setStatusBar(data: ColorMap) {
   data.set('statusBar.background', editorColorDef.accent);
   data.set('statusBar.foreground', editorColorDef.accentFg);
   data.set('statusBarItem.hoverBackground', editorColorDef.accentHover);