+++ src/semantic/variableSemantic.ts
@@ -1,0 +1,7 @@
+import { syntaxColors, tokenColorsDef } from '../color';
+
+export default function variableSemantic(data: ColorMap): void {
+  data.set('property', tokenColorsDef.default);
+
+  data.set('property.declaration', syntaxColors.cyan);
+}